feat: 缓存策略,允许entity实例控制自己是否需要缓存

+++ cache.go
@@ -32,6 +32,8 @@ type CacheOption struct {
 	Key        string
 	Expiration time.Duration
 	Compress   bool
+	// 如果设置为true，entity实例的缓存功能会被关闭
+	Disable bool
 	// 某些由其它地方构造的缓存，其中存在字段内容进入缓存前先被json encode过
 	// 这些字段缓存结果需要被decode两次才能使用
 	RecursiveDecode []string
@@ -41,6 +43,9 @@ func loadCache(ctx context.Context, ent Cacheable) (bool, error) {
 	opt, err := getCacheOption(ent)
 	if err != nil {
 		return false, fmt.Errorf("get option, %w", err)
+	}
+	if opt.Disable {
+		return false, nil
 	}
 
 	data, err := opt.Cacher.Get(ctx, opt.Key)
@@ -81,14 +86,17 @@ func loadCache(ctx context.Context, ent Cacheable) (bool, error) {
 
 // SaveCache 保存entity缓存
 func SaveCache(ctx context.Context, ent Cacheable) error {
-	data, err := jsoniter.Marshal(ent)
-	if err != nil {
-		return fmt.Errorf("json encode, %w", err)
-	}
-
 	opt, err := getCacheOption(ent)
 	if err != nil {
 		return fmt.Errorf("get option, %w", err)
+	}
+	if opt.Disable {
+		return nil
+	}
+
+	data, err := jsoniter.Marshal(ent)
+	if err != nil {
+		return fmt.Errorf("json encode, %w", err)
 	}
 
 	if opt.Compress {
@@ -111,6 +119,9 @@ func DeleteCache(ctx context.Context, ent Cacheable) error {
 	opt, err := getCacheOption(ent)
 	if err != nil {
 		return fmt.Errorf("get option, %w", err)
+	}
+	if opt.Disable {
+		return nil
 	}
 
 	return opt.Cacher.Delete(ctx, opt.Key)
